Migrate deprecations in test
